requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
pydantic>=2.5,<3
openai>=1.0,<2
httpx>=0.27,<1
python-dotenv

pytest>=8,<9
pytest-asyncio>=0.23,<1

